+++ include/periple/periple.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
 #include <periple/core/traits.hpp>
-#include <periple/core/callbacks.hpp>
-#include <periple/distance/matrix.hpp>
-#include <periple/core/jonker_volgenant.hpp>
+#include <periple/core/moves.hpp>
 #include <periple/core/solver.hpp>
+#include <periple/distance/coordinate.hpp>
+#include <periple/distance/functions.hpp>
+#include <periple/distance/jonker_volgenant.hpp>
+#include <periple/distance/matrix.hpp>
 #include <periple/algorithms/greedy_construct.hpp>
 #include <periple/algorithms/nearest_neighbor.hpp>
 #include <periple/algorithms/held_karp.hpp>