+++ tests/logging_callbacks.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <periple/core/callbacks.hpp>
+#include <periple/core/moves.hpp>
 
 #include <span>
 #include <string>